feat: create package.ts util [sc-00]
